requirements-ingest.txt
# Python dependencies for DOCX ingestion
pandas>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
